+++ src/components/Login.js
@@ -138,7 +138,7 @@ const Login = () => {
         style={{ height: '100%' }}
         verticalAlign='middle'>
         <Grid.Column style={{ maxWidth: 550 }}>
-          <Image size='big' centered src={require('./logo.png')} />
+          <Image size='big' centered src={require('./Misc/logo.png')} />
           <Header as='h2' color='teal' textAlign='center'>
               Log-in to your account
           </Header>
