+++ src/components/MainCtrl.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
 import { Container, Grid, Dimmer, Loader } from 'semantic-ui-react'
 import axios from 'axios'
-import Topbar from './Topbar'
-import SideMenu from './SideMenu'
+import Topbar from './Misc/Topbar'
+import SideMenu from './Misc/SideMenu'
 import ClassWrap from './Class/ClassWrap'
-import Home from './Home'
+import Home from './Home/Home'
 import { object } from 'prop-types'
 import VolunteerWrap from './Volunteer/VolunteerWrap'
 import AttendanceWrap from './Attendance/AttendanceWrap'
