+++ src/index.js
@@ -5,9 +5,9 @@ import MainCtrl from './components/MainCtrl'
 import StudentRegister from './components/Register/StudentRegistration/StudentRegisterForm'
 import VolunteerRegister from './components/Register/VolunteerRegistration/VolunteerRegisterForm'
 import Login from './components/Login'
-import ForgetPassword from './components/ForgetPassword'
-import ResetPassword from './components/ResetPassword'
-import EmailVerify from './components/EmailVerify'
+import ForgetPassword from './components/Misc/ForgetPassword'
+import ResetPassword from './components/Misc/ResetPassword'
+import EmailVerify from './components/Misc/EmailVerify'
 import Home from './components/Main/Home'
 import 'semantic-ui-css/semantic.min.css'
 import './index.css'
